+++ marrow/src/impl_arrow/mod.rs
@@ -1,11 +1,11 @@
 //! Support for arrow
 #![cfg_attr(any(), rustfmt::skip)]
 
-// arrow-version:insert: #[cfg(feature = "arrow-{version}")]{\n}{\n}mod arrow_{version} {{{\n}    use {{arrow_array_{version} as arrow_array, arrow_buffer_{version} as arrow_buffer, arrow_data_{version} as arrow_data, arrow_schema_{version} as arrow_schema}};{\n}    include!("impl_api_47.rs");{\n}}}
+// arrow-version:insert: #[cfg(feature = "arrow-{version}")]{\n}{\n}mod arrow_{version} {{{\n}    use {{arrow_array_{version} as arrow_array, arrow_buffer_{version} as arrow_buffer, arrow_data_{version} as arrow_data, arrow_schema_{version} as arrow_schema}};{\n}    include!("impl_api_53.rs");{\n}}}
 #[cfg(feature = "arrow-53")]
 mod arrow_53 {
     use {arrow_array_53 as arrow_array, arrow_buffer_53 as arrow_buffer, arrow_data_53 as arrow_data, arrow_schema_53 as arrow_schema};
-    include!("impl_api_47.rs");
+    include!("impl_api_53.rs");
 }
 #[cfg(feature = "arrow-52")]
 mod arrow_52 {